# src/src-vpn/docs/api.yaml
openapi: "3.0.3"
info:
  title: Unkillable Messenger API
  version: "1.0.0"
  description: REST API for the Unkillable Messenger platform.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/files/{id}:
    get:
      summary: Download a file
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File data
...
  /api/bots/register:
    post:
      summary: Register a bot
      operationId: registerBot
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                command: { type: string }
      responses:
        "200":
          description: Bot registered
...